+++ src/stores/derrivedYobta/index.ts
@@ -9,14 +9,14 @@ type States<Stores extends YobtaAnyStore[]> = {
     ? State
     : YobtaState<Stores[Key]>
 }
-interface YobtaDerrived {
+interface Yobtaderived {
   <
-    DerrivedState extends any = any,
+    derivedState extends any = any,
     Stores extends YobtaAnyStore[] = YobtaAnyStore[],
   >(
-    callback: (...states: States<Stores>) => DerrivedState,
+    callback: (...states: States<Stores>) => derivedState,
     ...stores: Stores
-  ): YobtaReadable<DerrivedState, never>
+  ): YobtaReadable<derivedState, never>
 }
 // #endregion
 
@@ -26,10 +26,10 @@ interface YobtaDerrived {
  * @example
  * const store1 = storeYobta(1)
  * const store2 = storeYobta(1)
- * const derrived = derrivedYobta((state1, state2) => state1 + state2, store1, store2)
- * @documentation {@link https://github.com/yobta/stores/tree/master/src/stores/derrivedYobta/index.md}
+ * const derived = derivedYobta((state1, state2) => state1 + state2, store1, store2)
+ * @documentation {@link https://github.com/yobta/stores/tree/master/src/stores/derivedYobta/index.md}
  */
-export const derrivedYobta: YobtaDerrived = (acc, ...stores) => {
+export const derivedYobta: Yobtaderived = (acc, ...stores) => {
   let getState = (): any =>
     acc(...(stores.map(({ last }) => last()) as States<typeof stores>))
   let { last, on, next, observe } = storeYobta<any, never>(getState())