+++ src/index.ts
@@ -1,3 +1,4 @@
+export * from './stores/derivedYobta/index.js'
 export * from './stores/machineYobta/index.js'
 export * from './stores/mapYobta/index.js'
 export * from './stores/onlineYobta/index.js'