+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.1.13
 
-Derrived store
+derived store
 
 ## 0.1.12